Crate pg_extern_attr

Attribute Macros

pg_extern
An attribute macro for wrapping Rust functions with boiler plate for defining and calling conventions between Postgres and Rust.

pg_foreignwrapper
An attribute macro for wrapping Rust structs with boiler plate for defining and exposing a foreign data wrapper This is mostly a slimmed down version of pg_extern, with none of the data argument handling.